Correct `FromOkm` for `DecafScalar`
