+++ ed448-goldilocks/src/curve/scalar.rs
@@ -1,9 +1,11 @@
 use crate::field::{CurveWithScalar, NZ_ORDER, Scalar, ScalarBytes, WideScalarBytes};
 use crate::{Ed448, ORDER};
 
-use elliptic_curve::bigint::{Limb, U448};
-use elliptic_curve::consts::U57;
+use elliptic_curve::array::Array;
+use elliptic_curve::bigint::{Limb, NonZero, U448, U704};
+use elliptic_curve::consts::{U57, U84, U88};
 use elliptic_curve::scalar::FromUintUnchecked;
+use hash2curve::FromOkm;
 use subtle::{Choice, CtOption};
 
 impl CurveWithScalar for Ed448 {
@@ -77,6 +79,24 @@ pub type WideEdwardsScalarBytes = WideScalarBytes<Ed448>;
 impl From<&EdwardsScalar> for elliptic_curve::scalar::ScalarBits<Ed448> {
     fn from(scalar: &EdwardsScalar) -> Self {
         scalar.scalar.to_words().into()
+    }
+}
+
+impl FromOkm for EdwardsScalar {
+    type Length = U84;
+
+    fn from_okm(data: &Array<u8, Self::Length>) -> Self {
+        const SEMI_WIDE_MODULUS: NonZero<U704> = NonZero::<U704>::new_unwrap(U704::from_be_hex(
+            "00000000000000000000000000000000000000000000000000000000000000003fffffffffffffffffffffffffffffffffffffffffffffffffffffff7cca23e9c44edb49aed63690216cc2728dc58f552378c292ab5844f3",
+        ));
+        let mut tmp = Array::<u8, U88>::default();
+        tmp[4..].copy_from_slice(&data[..]);
+
+        let mut num = U704::from_be_slice(&tmp[..]);
+        num %= SEMI_WIDE_MODULUS;
+        let mut words = [0; U448::LIMBS];
+        words.copy_from_slice(&num.to_words()[..U448::LIMBS]);
+        Scalar::new(U448::from_words(words))
     }
 }
